requirements.txt
Flask
Pillow
exifread
PyPDF2
python-docx
mutagen
python-magic-bin
pytest
